move `ui-*` components into `common`
