+++ src/client/modules/common/components/web/ui-antd/components/NavBar.jsx
@@ -5,8 +5,8 @@ import Menu from 'antd/lib/menu';
 import Row from 'antd/lib/row';
 import Col from 'antd/lib/col';
 
-import modules from '../../../modules';
-import settings from '../../../../../settings';
+import modules from '../../../../../../modules';
+import settings from '../../../../../../../../settings';
 
 class NavBar extends React.Component {
   state = {
