+++ src/client/modules/common/components/web/ui-antd/index.web.js
@@ -1,6 +1,6 @@
 import styles from './styles/styles.less';
 
-import Feature from '../connector';
+import Feature from '../../../../connector';
 
 export default new Feature({
   stylesInsert: styles
