+++ src/client/modules/common/components/web/ui-antd/styles/antd.less
@@ -1,0 +1,1 @@
+@import '../../../../../../../../node_modules/antd/dist/antd';
