+++ src/client/modules/common/components/web/ui-bootstrap/components/NavBar.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { NavLink } from 'react-router-dom';
 import { Container, Navbar, Nav, NavItem } from 'reactstrap';
 
-import modules from '../../../modules';
-import settings from '../../../../../settings';
+import modules from '../../../../../../modules';
+import settings from '../../../../../../../../settings';
 
 const NavBar = () => (
   <Navbar color="faded" light>
