+++ src/client/modules/common/components/web/ui-bootstrap/components/index.js
