+++ src/client/modules/common/components/web/ui-bootstrap/index.web.js
@@ -1,6 +1,6 @@
 import styles from './styles/styles.scss';
 
-import Feature from '../connector';
+import Feature from '../../../../connector';
 
 export default new Feature({
   stylesInsert: styles
