+++ src/client/modules/common/components/web/ui-bootstrap/styles/bootstrap.scss
@@ -1,0 +1,1 @@
+@import '../../../../../../../../node_modules/bootstrap/scss/bootstrap';
