--- src/client/modules/ui-antd/styles/antd.less
@@ -1,1 +1,0 @@
-@import '../../../../../node_modules/antd/dist/antd';
