--- src/client/modules/ui-bootstrap/styles/bootstrap.scss
@@ -1,1 +1,0 @@
-@import '../../../../../node_modules/bootstrap/scss/bootstrap';
